- new issues solved (#5)
* - new issues solved

* - added statistics

* - added how to use and minor fix

* - fixed one chart issue

* - timings stats issue fixed

* - added google apps

* - completed timing calculation

* - completed timing calculation - 2

* - varun sir issue completed except 1

* Refactor: Move app count increment to successful unlock

The `incrementAppCount` call for the `lockedPackageName` is now only executed when the app is successfully unlocked by entering the correct password. Previously, it was also called when the app was manually blocked.

Additionally, a debug print statement related to manual close database updates has been removed.

* Refactor: Move app count increment to successful unlock

The `incrementAppCount` call for the `lockedPackageName` is now only executed when the app is successfully unlocked by entering the correct password. Previously, it was also called when the app was manually blocked.

Additionally, a debug print statement related to manual close database updates has been removed.

* - assets added

---------

Co-authored-by: parthas <partha.s@zoozle.in>